refactor: Batter 와 Pitcher 가 Player 를 상속 받도록 함
