--- be/server/src/main/java/com/baseball/domain/InningResult.java
@@ -1,10 +1,0 @@
-package com.baseball.domain;
-
-public enum InningResult {
-    BALL, HIT, STRIKE;
-
-    public InningResult of(String name) {
-        name = name.toUpperCase();
-        return valueOf(name);
-    }
-}
