+++ be/server/src/main/java/com/baseball/domain/PitchResult.java
@@ -1,0 +1,14 @@
+package com.baseball.domain;
+
+public enum PitchResult {
+    BALL, HIT, STRIKE;
+
+    public PitchResult of(String name) {
+        name = name.toUpperCase();
+        return valueOf(name);
+    }
+
+    public Boolean toBoolean() {
+        return this == STRIKE;
+    }
+}
