+++ be/server/src/main/java/com/baseball/domain/player/Players.java
@@ -1,4 +1,4 @@
-package com.baseball.domain;
+package com.baseball.domain.player;
 
 import java.util.List;
